requirements.txt
discord.py==1.6.0
SQLAlchemy==1.3.23
